optimize rendering of children that have common model datasources with their direct parent

+++ src/Controller.js
@@ -110,6 +110,8 @@ define([
 
             var changeListenerDestroyFunctions = self._changeListenerDestroyFunctions = [];
 
+            self._modelKeys = {};
+
             datasources.forEach(function(source) {
                 pending.push(function() {
                     var sourceKeys = Object.keys(source),
@@ -130,12 +132,20 @@ define([
                     });
 
                     return CurvilinearPromise.parallelize(promises).then(function(results) {
-                        if (!pending.cancelled) {
+                        if (!pending.cancelled && !self._destroyed) {
                             results.forEach(function(result, i) {
                                 var modelKey = result instanceof ModelValue && result.key;
 
                                 if (typeof modelKey === 'string') {
-                                    changeListenerDestroyFunctions.push(model.observe(modelKey, self.render.bind(self)));
+                                    self._modelKeys[modelKey] = true;
+
+                                    if (!self._parent || !self._parent._modelKeys[modelKey]) {
+                                        changeListenerDestroyFunctions.push(model.observe(modelKey, function() {
+                                            if (!self._destroyed) {
+                                                self.render();
+                                            }
+                                        }));
+                                    }
 
                                     result = result.value;
                                 }
@@ -150,7 +160,7 @@ define([
             var mainPromise = new CurvilinearPromise();
 
             CurvilinearPromise.serialize(pending).then(function() {
-                if (!pending.cancelled) {
+                if (!pending.cancelled && !self._destroyed) {
                     self._data = Object.freeze(newData);
                     self._pending = null;
                     self._destroyChildren();
@@ -169,6 +179,8 @@ define([
                         var childPromises = new Array(children.length);
 
                         children.forEach(function(child, i) {
+                            child._parent = self;
+
                             childPromises[i] = child.render();
 
                             self._children.push(child);
